Improved and reorganised the scripts

+++ scripts/build.sh
@@ -12,7 +12,7 @@ LANGUAGES_FILE="languages.xml"
 LOCATIONS_FILE="locations.xml"
 TITLES_FILE="titles.xml"
 
-if [ -z "${BUILD_VERSION}" ] || ! [[ ${BUILD_VERSION} =~ '^[0-9]+$' ]]; then
+if [ -z "${BUILD_VERSION}" ] || ! [[ ${BUILD_VERSION} =~ ^[0-9]+$ ]]; then
     BUILD_VERSION=0
 fi
 
@@ -20,6 +20,7 @@ VERSION=$(date +"%y").$(date +"%j").${BUILD_VERSION}
 
 if [[ $* != *--skip-updates* ]]; then
     bash "${SCRIPTSDIR}/update-builder.sh"
+    bash "${SCRIPTSDIR}/update-vanilla-files.sh"
 fi
 
 if [[ $* != *--skip-validation* ]]; then
@@ -41,11 +42,11 @@ function build-edition {
     PACKAGE_NAME="mcn_${GAME}_${VERSION}"
     ORIGINAL_WORKING_DIRECTORY=$(pwd)
 
-    [ -d "${OUTDIR}/${GAME}" ] && rm -rf "${OUTDIR}/${GAME}"
+    [ -d "${OUTDIR}/${GAME}" ] && rm -rf "${OUTDIR:?}/${GAME:?}"
     [ -f "${OUTDIR}/${PACKAGE_NAME}.zip" ] && rm "${OUTDIR}/${PACKAGE_NAME}.zip"
 
     cd "${STARTDIR}"
-    "${STARTDIR}/more-cultural-names-builder/MoreCulturalNamesModBuilder" \
+    "${STARTDIR}/.builder/MoreCulturalNamesModBuilder" \
         --lang "${LANGUAGES_FILE}" \
         --loc "${LOCATIONS_FILE}" \
         --titles "${TITLES_FILE}" \
@@ -55,7 +56,7 @@ function build-edition {
 
     if [ ! -d "${OUTDIR}/${GAME}/" ]; then
         echo "   > ERROR: Failed to build the ${GAME} edition!"
-        exit 400
+        exit 200
     fi
 
     echo "   > Copying extras..."
@@ -72,33 +73,33 @@ function build-edition {
 build-edition \
     "more-cultural-names" "More Cultural Names" \
     "CK2" "3.3.5.1" \
-    --landed-titles "vanilla//ck2_landed_titles.txt" --landed-titles-name "landed_titles.txt"
+    --landed-titles "${VANILLA_FILES_DIR}/ck2_landed_titles.txt" --landed-titles-name "landed_titles.txt"
 
 build-edition \
     "hip-more-cultural-names" "HIP - More Cultural Names" \
     "CK2HIP" "Frosty3" \
-    --landed-titles "vanilla/ck2hip_landed_titles.txt" --landed-titles-name "swmh_landed_titles.txt" \
+    --landed-titles "${VANILLA_FILES_DIR}/ck2hip_landed_titles.txt" --landed-titles-name "swmh_landed_titles.txt" \
     --dep "HIP - Historical Immersion Project"
 
 build-edition \
     "more-cultural-names" "More Cultural Names" \
     "CK3" "1.4.*" \
-    --landed-titles "vanilla/ck3_landed_titles.txt" --landed-titles-name "999_MoreCulturalNames.txt"
+    --landed-titles "${VANILLA_FILES_DIR}/ck3_landed_titles.txt" --landed-titles-name "999_MoreCulturalNames.txt"
 
 build-edition \
     "ibl-more-cultural-names" "Ibn Battuta's Legacy 2 - More Cultural Names" \
     "CK3IBL" "1.4.*" \
-    --landed-titles "vanilla/ck3ibl_landed_titles.txt" --landed-titles-name "873_MoreCulturalNames.txt"
+    --landed-titles "${VANILLA_FILES_DIR}/ck3ibl_landed_titles.txt" --landed-titles-name "873_MoreCulturalNames.txt"
 
 build-edition \
     "mbp-more-cultural-names" "More Bookmarks+ - More Cultural Names" \
     "CK3MBP" "1.4.*" \
-    --landed-titles "vanilla/ck3mbp_landed_titles.txt" --landed-titles-name "873_MoreCulturalNames.txt"
+    --landed-titles "${VANILLA_FILES_DIR}/ck3mbp_landed_titles.txt" --landed-titles-name "873_MoreCulturalNames.txt"
 
 build-edition \
     "tfe-more-cultural-names" "The Fallen Eagle: More Cultural Names" \
     "CK3TFE" "1.4.*" \
-    --landed-titles "vanilla/ck3tfe_landed_titles.txt" --landed-titles-name "873_MoreCulturalNames.txt"
+    --landed-titles "${VANILLA_FILES_DIR}/ck3tfe_landed_titles.txt" --landed-titles-name "873_MoreCulturalNames.txt"
 
 build-edition \
     "more-cultural-names" "More Cultural Names" \
